Mis à jour des fichiers *.in

+++ src/meshfem2D/scotch_6.0.4/include/scotch.h
@@ -73,7 +73,7 @@
 
 /*+ Integer type. +*/
 
-typedef int64_t SCOTCH_Idx;
+typedef int SCOTCH_Idx;
 
 typedef int SCOTCH_Num;
